pyproject.toml
[project]
name = "github-profile-finder"
version = "0.1.0"
description = "Find promising developers from GitHub repositories they've forked"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    {name = "Odysseas Lamtzidis", email = "odyslam@gmail.com"}
]
keywords = ["github", "developer", "profiles", "recruiting", "talent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "requests>=2.28.1",
    "tabulate>=0.9.0",
    "tomli>=2.0.1",
    "tomli-w>=1.0.0",
    "python-dotenv>=1.0.0",
    "pygithub>=2.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/odyslam/gh-finder"
Issues = "https://github.com/odyslam/gh-finder/issues"
Documentation = "https://github.com/odyslam/gh-finder/blob/main/README.md"

[project.scripts]
gh-finder = "gh_finder.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gh_finder"]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
